[refactor] useQueryInProgressBiddings goes to atom
Motivation is to cut 1 firestore listener, since the entire collection
is already listened to, no need to listen for the query which is a
subset of the collection.

Alternative is to not use Atom but use the async state management of
tanstack query. That would keep the useQuery removed in this
commit. But it involves using setQuery to to update the firebase query
at a very ugly queryKey. Not to mention is would be async unnecessarily.